Track which bracket transitions can happen simultaneously.
In order to fully determinize an automaton that includes bracket
transitions, we need to know what the resulting bracket automaton's
accepting states will be.  Since the bracket automaton itself can
include bracket transitions, we have to keep determinizing until we reach
a fixed point (we're guaranteed to reach it due to monotonicity).