Stop trying to reverse-engineer determinized bracket symbols.
Two different states with the same symbol is bad news, but we can end up
in this situation when two different states have the same bracket symbols
(that's all we match against when we're reverse-engineering with
MARK_ACCEPTING_BRACKET_STATES).

Instead, remove MARK_ACCEPTING_BRACKET_STATES and just do the work as
part of `find_bracket_transitions`.  Make sure we clear the action map,
since we'll be writing to it over and over while iteratively finding
bracket transitions.