Stop writing to a closed FILE.
The write_to_output function writes to output_file, but during compilation
tests, output_file is closed.  This worked by accident, probably because
the pipe used to send the generated code to the compiler happened to get
the same file descriptor that the output file had.

The fix is to set output_file to test.file directly.